+++ src/main/c/seasocks/Connection.h
@@ -176,10 +176,10 @@ private:
     void error(ResponseCode responseCode, const std::string& payload);
 
     struct Range {
-        long start;
-        long end;
+        size_t start;
+        size_t end;
         size_t length() const {
-            return (size_t) ((size_t) end - size_t(start) + size_t(1));
+            return end - start + 1;
         }
     };
 
@@ -191,7 +191,7 @@ private:
 
     void bufferResponseAndCommonHeaders(ResponseCode code);
 
-    std::list<Range> processRangesForStaticData(const std::list<Range>& ranges, long fileSize);
+    std::list<Range> processRangesForStaticData(const std::list<Range>& ranges, size_t fileSize);
 
     std::shared_ptr<Logger> _logger;
     ServerImpl& _server;
